requirements.txt
matplotlib==3.4.1
numpy==1.19.2
pandas
tensorflow
google-cloud-storage==1.37.1
opencv-python
SciPy